requirements.txt
adafruit-circuitpython-epd
adafruit-blinka
adafruit-platformdetect
pillow
requests
wheel
adafruit-circuitpython-minimqtt
